ocr-service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
# Pin <3.3.0 to avoid CPU inference bug: ConvertPirAttribute2RuntimeAttribute not support (oneDNN). See PaddlePaddle/Paddle#77340
paddlepaddle>=2.5.0,<3.3.0
paddleocr>=2.7.0
pdf2image>=1.16.0
Pillow>=10.0.0
python-multipart>=0.0.6
openai>=1.40.0
